feat(Assertions): refactor assertions, combine React's to single
