--- src/util/getBody.js
@@ -1,3 +1,0 @@
-const getBody = ({ node: { body } }) => body.body
-
-export default getBody
