+++ src/util/index.js
@@ -1,19 +1,16 @@
-export { default as createPropertyExpression } from './createPropertyExpression'
-export { default as insertAfterPath } from './insertAfterPath'
+export { default as appendProps } from './appendProps'
 
 export { default as getClassDeclaration } from './getClassDeclaration'
 export { default as getExpressionIdentifier } from './getExpressionIdentifier'
 export { default as getEntryIdentifier } from './getEntryIdentifier'
 
-export { default as isReactClass } from './isReactClass'
-export { default as isReactFunction } from './isReactFunction'
+export { default as isReactComponent } from './isReactComponent'
 export { default as isReactImport } from './isReactImport'
 
 export { default as Store } from './Store'
 
 export {
   isArrayValue,
-  isObjectExpression,
   isObjectValue,
   isStaticProperty,
   isValidExpression,
