+++ src/visitors/entryVisitor.js
@@ -1,12 +1,8 @@
-import {
-  getEntryIdentifier,
-  isReactClass,
-  isReactFunction,
-} from '../util'
+import { getEntryIdentifier, isReactComponent } from '../util'
 
 const entriesVisitor = {
   'Class|Function'(path, state) {
-    if (!isReactClass(path) && !isReactFunction(path)) return
+    if (!isReactComponent(path)) return
     state.createEntry(getEntryIdentifier(path), path)
   },
 }
