+++ test/fixtures/multiple/expected.js
@@ -1,12 +1,12 @@
 import React, { PropTypes } from 'react';
 
 export function First() {
-  return null;
+  return React.createElement('div', null);
 }
 
 First.handledProps = ['children'];
 export function Second() {
-  return null;
+  return React.createElement('div', null);
 }
 
 Second.handledProps = ['children'];
