+++ test/fixtures/stateless-assignment/expected.js
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 
 const Example = function () {
-  return null;
+  return React.createElement('div', null);
 };
 Example.handledProps = ['active', 'children', 'className'];
 Example.defaultProps = {
